Add a script to make a bunch of test. Can now set the timeout and provide options files for the stochastic, generation, and evaluation solvers.

+++ src/Solver.h
@@ -311,7 +311,7 @@ class SolverParam{
 public:
 	SolverParam(){}
    //maximal solving time in s
-   int maxSolvingTimeSeconds_ = LARGE_TIME;
+   double maxSolvingTimeSeconds_ = LARGE_TIME;
 
    //print parameters
    bool printEverySolution_ = false;
