+++ src/StochasticSolver.h
@@ -60,7 +60,7 @@ public:
 	// RK_MEAN: keep the schedule with minimum expected cost over the generated evaluation demands
 	RankingStrategy rankingStrategy_ = RK_SCORE;
 	bool demandingEvaluation_ = true;
-	int totalTimeLimitSeconds_ = LARGE_TIME;
+	double totalTimeLimitSeconds_ = LARGE_TIME;
 
 	// Number of evaluation demands generated
 	// WARNING: if =0 and withEvaluation_=true, ranks the schedules according to their baseCost (i.e. the "real" cost of the 1-week schedule [without min/max costs])
